Fix a bug when swapfile doesn't exist

+++ terraform-aws-github-runner/modules/runners-instances/templates/user-data.sh
@@ -138,9 +138,11 @@ fi
 
 ${post_install}
 
-# Cleanup any existing swapfile just to be sure
-sudo swapoff /swapfile
-sudo rm /swapfile
+if [[ -f /swapfile ]]; then
+  # Cleanup any existing swapfile just to be sure
+  sudo swapoff /swapfile
+  sudo rm /swapfile
+fi
 # before allocating a new one
 sudo fallocate -l 3G /swapfile
 sudo chmod 600 /swapfile
